+++ app/src/androidTest/java/org/simple/clinic/patient/PatientRepositoryAndroidTest.kt
@@ -478,7 +478,7 @@ class PatientRepositoryAndroidTest {
 
     val patient = patientProfile.patient
 
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val updatedAfter = Duration.ofDays(1L)
     clock.advanceBy(updatedAfter)
@@ -541,7 +541,7 @@ class PatientRepositoryAndroidTest {
         businessIds = emptyList()
     )
 
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val updatedAfter = Duration.ofDays(1L)
     clock.advanceBy(updatedAfter)
@@ -598,7 +598,6 @@ class PatientRepositoryAndroidTest {
     )
 
     patientRepository.save(listOf(patientProfile))
-        .blockingAwait()
 
     val updatedAfter = Duration.ofDays(1L)
     clock.advanceBy(updatedAfter)
@@ -649,7 +648,6 @@ class PatientRepositoryAndroidTest {
         businessIds = emptyList())
 
     patientRepository.save(listOf(patientProfile))
-        .blockingAwait()
 
     val updatedAfter = Duration.ofDays(1L)
     clock.advanceBy(updatedAfter)
@@ -727,7 +725,7 @@ class PatientRepositoryAndroidTest {
         appointmentType = Manual,
         cancelReason = null
     )
-    appointmentRepository.save(listOf(appointment2)).blockingAwait()
+    appointmentRepository.save(listOf(appointment2))
 
     recentPatient2 = recentPatient2.copy(updatedAt = clock.instant())
     verifyRecentPatientOrder(
@@ -842,7 +840,7 @@ class PatientRepositoryAndroidTest {
     val patientProfile = testData.patientProfile(patientUuid = patientUuid, patientAddressUuid = patientAddressUuid).run {
       copy(patient = patient.copy(status = patientStatus))
     }
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val bpMeasurement = testData.bloodPressureMeasurement(
         patientUuid = patientUuid,
@@ -867,7 +865,7 @@ class PatientRepositoryAndroidTest {
     val patientProfile = testData.patientProfile(patientUuid = patientUuid).run {
       copy(patient = patient.copy(createdAt = createdAt, updatedAt = updatedAt))
     }
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val bpMeasurement = testData.bloodPressureMeasurement(
         patientUuid = patientUuid,
@@ -892,7 +890,7 @@ class PatientRepositoryAndroidTest {
     val patientProfile = testData.patientProfile(patientUuid = patientUuid).run {
       copy(patient = patient.copy(createdAt = createdAt, updatedAt = updatedAt))
     }
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val bloodSugarMeasurement = testData.bloodSugarMeasurement(
         patientUuid = patientUuid,
@@ -949,7 +947,7 @@ class PatientRepositoryAndroidTest {
       deletedAt: Instant? = null
   ): RecentPatient {
     val patientProfile = testData.patientProfile(patientUuid = patientUuid)
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val prescribedDrug = testData.prescription(
         patientUuid = patientUuid,
@@ -1086,7 +1084,7 @@ class PatientRepositoryAndroidTest {
       appointmentType: AppointmentType = Manual
   ): RecentPatient {
     val patientProfile = testData.patientProfile(patientUuid = patientUuid)
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val appointment = testData.appointment(
         uuid = appointmentUuid,
@@ -1118,7 +1116,7 @@ class PatientRepositoryAndroidTest {
         testData.patientProfile(syncStatus = PENDING),
         testData.patientProfile(syncStatus = PENDING),
         testData.patientProfile(syncStatus = DONE)
-    )).blockingAwait()
+    ))
 
     val count = patientRepository.pendingSyncRecordCount().blockingFirst()
     assertThat(count).isEqualTo(2)
@@ -1127,7 +1125,7 @@ class PatientRepositoryAndroidTest {
   @Test
   fun saving_a_bp_identifier_for_a_patient_must_work_as_expected() {
     val patientProfile = testData.patientProfile(syncStatus = DONE, generateBusinessId = false)
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val bpPassportCode = "6adbfe88-840e-4f03-aaaf-4b1f2bae747a"
     val now = Instant.now(clock)
@@ -1167,7 +1165,7 @@ class PatientRepositoryAndroidTest {
   @Test
   fun saving_bangladesh_national_identifier_for_a_patient_must_work_as_expected() {
     val patientProfile = testData.patientProfile(syncStatus = DONE, generateBusinessId = false)
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val nationalId = "33ed3fb2-cfcc-48f8-9b7d-079c02146076"
     val now = Instant.now(clock)
@@ -1387,8 +1385,7 @@ class PatientRepositoryAndroidTest {
         patientWithUniqueBusinessId,
         patientOneWithSharedBusinessId,
         patientTwoWithSharedBusinessId,
-        patientWithDeletedBusinessId)
-    ).blockingAwait()
+        patientWithDeletedBusinessId))
 
     val patientResultOne = patientRepository.findPatientWithBusinessId(identifier = uniqueBusinessIdentifier).blockingFirst().get()
     assertThat(patientResultOne).isEqualTo(patientWithUniqueBusinessId.patient)
@@ -1440,21 +1437,21 @@ class PatientRepositoryAndroidTest {
             systolic = it.reading.systolic,
             diastolic = it.reading.diastolic,
             recordedAt = it.recordedAt
-        ))).blockingAwait()
+        )))
       }
       bloodSugarMeasurement?.forEach {
         bloodSugarRepository.save(listOf(testData.bloodSugarMeasurement(
             patientUuid = patientUuid,
             reading = it.reading,
             recordedAt = it.recordedAt
-        ))).blockingAwait()
+        )))
       }
       medicalHistoryRepository.save(listOf(testData.medicalHistory(
           patientUuid = patientUuid,
           hasDiabetes = hasDiabetes,
           hasHadHeartAttack = hasHadHeartAttack,
           hasHadKidneyDisease = hasHadKidneyDisease,
-          hasHadStroke = hasHadStroke))).blockingAwait()
+          hasHadStroke = hasHadStroke)))
 
       protocolDrug?.let {
         prescriptionRepository.savePrescription(
@@ -1776,7 +1773,7 @@ class PatientRepositoryAndroidTest {
     val dummyProfile = testData.patientProfile(patientUuid = patientUuid, generateBusinessId = false)
     val profileWithBusinessIds = dummyProfile.copy(businessIds = listOf(oldBpPassport, currentBpPassport, deletedBpPassport))
 
-    patientRepository.save(listOf(profileWithBusinessIds)).blockingAwait()
+    patientRepository.save(listOf(profileWithBusinessIds))
 
     val latestBpPassport = patientRepository.bpPassportForPatient(patientUuid)
 
@@ -1809,7 +1806,7 @@ class PatientRepositoryAndroidTest {
     val recordedAtForPatient2 = Instant.now(clock)
     val patientProfile2 = createPatientProfile(patientUuid2, recordedAtForPatient2)
 
-    patientRepository.save(listOf(patientProfile1, patientProfile2)).blockingAwait()
+    patientRepository.save(listOf(patientProfile1, patientProfile2))
 
     patientRepository.compareAndUpdateRecordedAt(patientUuid1, instantToCompare1)
     patientRepository.compareAndUpdateRecordedAt(patientUuid2, instantToCompare2)
@@ -1851,7 +1848,7 @@ class PatientRepositoryAndroidTest {
     val recordedAtForPatient2 = Instant.now(clock)
     val patientProfile2 = createPatientProfile(patientUuid2, recordedAtForPatient2)
 
-    patientRepository.save(listOf(patientProfile1, patientProfile2)).blockingAwait()
+    patientRepository.save(listOf(patientProfile1, patientProfile2))
 
     patientRepository.compareAndUpdateRecordedAt(patientUuid1, instantToCompare1)
     patientRepository.compareAndUpdateRecordedAt(patientUuid2, instantToCompare2)
@@ -1882,7 +1879,7 @@ class PatientRepositoryAndroidTest {
               syncStatus = DONE)
           )
         }
-    patientRepository.save(listOf(patient)).blockingAwait()
+    patientRepository.save(listOf(patient))
 
     val bpRecordedTwoDaysBeforePatientCreated = testData.bloodPressureMeasurement(
         uuid = UUID.fromString("d24e9033-a397-4468-8a0a-4efb38c4dca1"),
@@ -1907,7 +1904,7 @@ class PatientRepositoryAndroidTest {
         bpRecordedTwoDaysBeforePatientCreated,
         bpRecordedOneDayBeforePatientCreated,
         bpRecordedOneMinuteAfterPatientCreated
-    )).blockingAwait()
+    ))
     patientRepository.updateRecordedAt(patientUuid).blockingAwait()
     loadPatient(patientUuid).let { savedPatient ->
       val expectedRecordedAt = bpRecordedTwoDaysBeforePatientCreated.recordedAt
@@ -1952,7 +1949,7 @@ class PatientRepositoryAndroidTest {
     }
     val patientUuid = patientProfile.patient.uuid
 
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val oneSecondAfterPatientUpdated = patientUpdatedAt.plus(Duration.ofSeconds(1L))
     val oneSecondBeforePatientUpdated = patientUpdatedAt.minus(Duration.ofSeconds(1L))
@@ -2053,7 +2050,7 @@ class PatientRepositoryAndroidTest {
           patientDeletedAt = if (isDeleted) Instant.now() else null
       )
 
-      patientRepository.save(listOf(patientProfile)).blockingAwait()
+      patientRepository.save(listOf(patientProfile))
     }
 
     fun findPatientWithIdentifier(identifier: String): Optional<Patient> {
@@ -2358,7 +2355,7 @@ class PatientRepositoryAndroidTest {
           zone = zone
       ))
     }
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     val patients = patientRepository.recordsWithSyncStatus(PENDING)
     assertThat(patients.size).isEqualTo(1)
@@ -2395,7 +2392,7 @@ class PatientRepositoryAndroidTest {
       )
     }
 
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     // when
     val patients = patientRepository.recordsWithSyncStatus(PENDING)
@@ -2414,7 +2411,7 @@ class PatientRepositoryAndroidTest {
         identifier = Identifier("12491236-9725-4eff-aab9-ba789ba37d72", BangladeshNationalId),
         createdAt = now
     )
-    patientRepository.save(listOf(patientProfile.copy(businessIds = listOf(bangladeshNationalId)))).blockingAwait()
+    patientRepository.save(listOf(patientProfile.copy(businessIds = listOf(bangladeshNationalId))))
 
     //when
     clock.advanceBy(Duration.ofHours(1))
@@ -2684,7 +2681,7 @@ class PatientRepositoryAndroidTest {
     )
     val expectedDeletedReason = DeletedReason.Duplicate
 
-    patientRepository.save(listOf(patientProfile)).blockingAwait()
+    patientRepository.save(listOf(patientProfile))
 
     // when
     clock.advanceBy(Duration.ofHours(1))
@@ -2719,7 +2716,7 @@ class PatientRepositoryAndroidTest {
 
     val updatedAssignedFacilityUuid = UUID.fromString("8fcdf153-d85f-4bcb-8cd3-b59ad5b1797a")
 
-    patientRepository.save(listOf(patient)).blockingAwait()
+    patientRepository.save(listOf(patient))
 
     // when
     clock.advanceBy(Duration.ofMinutes(40))
@@ -2769,7 +2766,7 @@ class PatientRepositoryAndroidTest {
         businessId = businessIdTwo
     )
 
-    patientRepository.save(listOf(patientOneProfile, patientTwoProfile)).blockingAwait()
+    patientRepository.save(listOf(patientOneProfile, patientTwoProfile))
 
     // when
     val patients = patientRepository.findPatientsWithBusinessId(identifier.value).map { it.uuid }
@@ -2797,7 +2794,7 @@ class PatientRepositoryAndroidTest {
               businessIds = businessIds
           )
 
-      patientRepository.save(listOf(patientProfile)).blockingAwait()
+      patientRepository.save(listOf(patientProfile))
     }
 
     fun searchResults(facility: Facility): List<String> {
@@ -2899,7 +2896,7 @@ class PatientRepositoryAndroidTest {
               businessIds = businessIds
           )
 
-      patientRepository.save(listOf(patientProfile)).blockingAwait()
+      patientRepository.save(listOf(patientProfile))
     }
 
     fun searchResults(query: String, facility: Facility): List<String> {
@@ -3005,7 +3002,7 @@ class PatientRepositoryAndroidTest {
               businessIds = businessIds
           )
 
-      patientRepository.save(listOf(patientProfile)).blockingAwait()
+      patientRepository.save(listOf(patientProfile))
     }
 
     fun searchResults(query: String, facility: Facility): List<String> {