+++ app/src/androidTest/java/org/simple/clinic/sync/BloodPressureSyncIntegrationTest.kt
@@ -114,7 +114,7 @@ class BloodPressureSyncIntegrationTest {
     }
     assertThat(records).containsNoDuplicates()
 
-    repository.save(records).blockingAwait()
+    repository.save(records)
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(totalNumberOfRecords)
 
     // when
@@ -142,12 +142,12 @@ class BloodPressureSyncIntegrationTest {
     }
     assertThat(records).containsNoDuplicates()
 
-    repository.save(records).blockingAwait()
+    repository.save(records)
     sync.push()
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(0)
 
     val modifiedRecord = records[1].withSystolic(250)
-    repository.save(listOf(modifiedRecord)).blockingAwait()
+    repository.save(listOf(modifiedRecord))
     assertThat(repository.pendingSyncRecordCount().blockingFirst()).isEqualTo(1)
 
     // when