+++ app/src/main/java/org/simple/clinic/drugs/search/DrugRepository.kt
@@ -1,7 +1,6 @@
 package org.simple.clinic.drugs.search
 
 import androidx.paging.PagingSource
-import io.reactivex.Completable
 import io.reactivex.Observable
 import org.simple.clinic.drugs.search.sync.DrugPayload
 import org.simple.clinic.patient.SyncStatus
@@ -13,8 +12,8 @@ class DrugRepository @Inject constructor(
     private val drugDao: Drug.RoomDao
 ) : SynceableRepository<Drug, DrugPayload> {
 
-  override fun save(records: List<Drug>): Completable {
-    return Completable.fromAction { saveRecords(records) }
+  override fun save(records: List<Drug>) {
+    saveRecords(records)
   }
 
   override fun setSyncStatus(from: SyncStatus, to: SyncStatus) {