+++ app/src/main/java/org/simple/clinic/sync/SynceableRepository.kt
@@ -1,6 +1,5 @@
 package org.simple.clinic.sync
 
-import io.reactivex.Completable
 import io.reactivex.Observable
 import org.simple.clinic.patient.SyncStatus
 import java.util.UUID
@@ -11,7 +10,7 @@ import java.util.UUID
  */
 interface SynceableRepository<T, P> {
 
-  fun save(records: List<T>): Completable
+  fun save(records: List<T>)
 
   fun setSyncStatus(from: SyncStatus, to: SyncStatus)