+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add a progress state in `CustomDrugEntrySheet`
 - Implement showing drug frequency in `EditMedicineScreen` based on the country
 - Add integration tests for `AppConfigRepository`
+- Change `SynceableRepository#save()` to be a synchronous call
 - [In Progress: 14 Sep 2021] Record call results instead of updating the same appointment record
 - [In Progress: 31 Aug 2021] Refactor logic around providing drug frequencies label depending on the country
 - [In Progress: 2 Sep 2021] Add a progress state in `CustomDrugEntrySheet`